feat: rename updatebonnesboites
