+++ server/src/jobs/lbb/updateOpcoCompanies.ts
@@ -5,7 +5,7 @@ import { logger } from "../../common/logger.js"
 import { Opco } from "../../common/model/index.js"
 import { logMessage } from "../../common/utils/logMessage.js"
 import { CFADOCK_FILTER_LIMIT, fetchOpcosFromCFADock } from "../../services/cfadock.service.js"
-import { checkIfAlgoFileIsNew, downloadAlgoCompanyFile, readCompaniesFromJson, removePredictionFile } from "./bonnesBoitesUtils.js"
+import { checkIfAlgoFileIsNew, downloadAlgoCompanyFile, readCompaniesFromJson, removePredictionFile } from "./lbaCompaniesUtils.js"
 import { getMemoizedOpcoShortName, saveOpco } from "../../services/opco.service.js"
 let errorCount = 0
